+++ src/views/Settings.vue
@@ -33,12 +33,7 @@
 					<v-expansion-panel-content>
 						<v-row class="mb-5">
 							<v-col cols="12" sm="6">
-								<v-switch
-									hint="Enable dark mode"
-									persistent-hint
-									label="Dark mode"
-									v-model="internalDarkMode"
-								></v-switch>
+								<color-scheme />
 							</v-col>
 							<v-col cols="12" sm="6">
 								<v-switch
@@ -804,31 +799,31 @@
 											<v-select
 												label="RF Region"
 												persistent-hint
-												hint="Will be applied on every startup if the current region of your Z-Wave controller differs. Leave this empty to use the default region of your stick. Not all controllers support changing the region."
+												hint="Will be used to automatically configure your controller for the region you are in. Not all controllers support changing the RF region."
 												:items="rfRegions"
-												clearable
+												:rules="[rules.required]"
+												required
 												v-model="newZwave.rf.region"
 											>
 											</v-select>
 										</v-col>
 										<v-col cols="6">
-											<v-select
-												label="Maximum LR Power Level"
+											<v-switch
+												hint="When enabled, both normal and LR power levels will be automatically set to legal limits based on the RF region whenever the region is changed. Only supported for Europe and USA regions."
 												persistent-hint
-												hint="The maximum power level to be used by the dynamic power algorithm of Z-Wave LR. Will be applied on every startup if the current setting of your Z-Wave controller differs. Only LR-capable controllers support this setting."
-												:items="maxLRPowerLevels"
-												clearable
+												label="Automatic Power Level"
 												v-model="
-													newZwave.rf
-														.maxLongRangePowerlevel
+													newZwave.rf.autoPowerlevels
 												"
-											>
-											</v-select>
+											></v-switch>
 										</v-col>
 									</v-row>
 									<v-row class="mt-0">
 										<v-col cols="12" sm="6">
 											<v-text-field
+												v-if="
+													!newZwave.rf.autoPowerlevels
+												"
 												label="Normal Power Level"
 												v-model.number="
 													newZwave.rf.txPower
@@ -860,6 +855,24 @@
 												type="number"
 												:rules="[validTxPower]"
 											></v-text-field>
+										</v-col>
+
+										<v-col cols="12" sm="6">
+											<v-select
+												label="Maximum LR Power Level"
+												v-if="
+													!newZwave.rf.autoPowerlevels
+												"
+												persistent-hint
+												hint="The maximum power level to be used by the dynamic power algorithm of Z-Wave LR. Will be applied on every startup if the current setting of your Z-Wave controller differs. Only LR-capable controllers support this setting."
+												:items="maxLRPowerLevels"
+												clearable
+												v-model="
+													newZwave.rf
+														.maxLongRangePowerlevel
+												"
+											>
+											</v-select>
 										</v-col>
 									</v-row>
 									<!-- END: RADIO CONFIGURATION -->
@@ -2030,7 +2043,7 @@
 			space-be
 			class="sticky-buttons py-3 px-4"
 			:style="{
-				backgroundColor: internalDarkMode ? '#272727' : '#f5f5f5',
+				backgroundColor: darkMode ? '#272727' : '#f5f5f5',
 			}"
 		>
 			<v-btn class="mr-2" small color="error" @click="resetConfig">
@@ -2072,7 +2085,11 @@ import { mapActions, mapState } from 'pinia'
 import ConfigApis from '@/apis/ConfigApis'
 import { parse } from 'native-url'
 import { wait, copy, isUndef, deepEqual } from '../lib/utils'
-import { rfRegions, znifferRegions, maxLRPowerLevels } from '../lib/items'
+import {
+	settingsRfRegions,
+	znifferRegions,
+	maxLRPowerLevels,
+} from '../lib/items'
 import cronstrue from 'cronstrue'
 import useBaseStore from '../stores/base'
 
@@ -2085,6 +2102,7 @@ export default {
 	name: 'Settings',
 	mixins: [InstancesMixin],
 	components: {
+		ColorScheme: () => import('@/components/custom/ColorScheme.vue'),
 		DialogGatewayValue: () =>
 			import('@/components/dialogs/DialogGatewayValue.vue'),
 		fileInput: () => import('@/components/custom/file-input.vue'),
@@ -2098,12 +2116,12 @@ export default {
 		},
 	},
 	computed: {
-		internalDarkMode: {
+		internalColorScheme: {
 			get() {
-				return this.darkMode
+				return this.colorScheme
 			},
 			set(value) {
-				this.setDarkMode(value)
+				this.setColorScheme(value)
 			},
 		},
 		internalNavTabs: {
@@ -2213,7 +2231,8 @@ export default {
 			'ui',
 		]),
 		...mapState(useBaseStore, {
-			darkMode: (store) => store.ui.darkMode,
+			colorScheme: (store) => store.ui.colorScheme,
+			darkMode: (store) => store.uiState.darkMode,
 			navTabs: (store) => store.ui.navTabs,
 			streamerMode: (store) => store.ui.streamerMode,
 		}),
@@ -2225,7 +2244,7 @@ export default {
 	},
 	data() {
 		return {
-			rfRegions,
+			rfRegions: settingsRfRegions,
 			znifferRegions,
 			maxLRPowerLevels,
 			valid_zwave: true,
@@ -2369,7 +2388,7 @@ export default {
 	},
 	methods: {
 		...mapActions(useBaseStore, [
-			'setDarkMode',
+			'setColorScheme',
 			'setNavTabs',
 			'setStreamerMode',
 			'initSettings',
@@ -2385,22 +2404,25 @@ export default {
 		validTxPower() {
 			const { powerlevel, measured0dBm } = this.newZwave.rf?.txPower ?? {}
 
-			const validPower = !isUndef(powerlevel)
-			const validMeasured = !isUndef(measured0dBm)
-
-			if (validPower && (powerlevel < -10 || powerlevel > 20)) {
-				return 'Power level must be between -10 and 20'
+			// The calibration value is optional
+			if (
+				typeof measured0dBm === 'number' &&
+				(measured0dBm < -10 || measured0dBm > 10)
+			) {
+				return 'Measured @ 0dBm must be between -10 and 10'
 			}
 
-			if (validMeasured && (measured0dBm < -10 || measured0dBm > 10)) {
-				return 'Measured 0dBm must be between -10 and 10'
+			if (isUndef(powerlevel) || powerlevel === 'auto') {
+				return
 			}
 
-			return (
-				(validPower && validMeasured) ||
-				(!validPower && !validMeasured) ||
-				'Both powerlevel and measured 0 dBm must be set when using custom TX power'
-			)
+			if (
+				typeof powerlevel !== 'number' ||
+				powerlevel < -10 ||
+				powerlevel > 20
+			) {
+				return 'Powerlevel must be between -10 and 20'
+			}
 		},
 		parseCron(cron) {
 			let res
@@ -2678,7 +2700,7 @@ export default {
 			this.newBackup = copy(this.backup)
 
 			if (this.prevUi) {
-				this.internalDarkMode = this.prevUi.darkMode
+				this.internalColorScheme = this.prevUi.colorScheme
 				this.internalNavTabs = this.prevUi.navTabs
 				this.internalStreamerMode = this.prevUi.streamerMode
 			} else {